requirements.txt
# core
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
joblib>=1.3.0
pyyaml>=6.0.1

# image + embeddings
pillow>=10.0.0
tqdm>=4.66.0

# pytorch (CPU)
torch>=2.2.0
torchvision>=0.17.0

# tests
pytest>=8.0.0
coverage>=7.0.0

# api
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.22
redis>=5,<8

# dvc
dvc>=3.0.0
dvc-gdrive>=3.0.0
fsspec<2025.7.0